Revert "create-chage-like-status"
This reverts commit 58ec7dc124c36848b89b625b1b8b3f0e5015eae3.
