--- src/main/java/com/bugflix/weblog/like/LikeController.java
@@ -1,24 +1,0 @@
-package com.bugflix.weblog.like;
-
-import com.bugflix.weblog.like.dto.LikeStatus;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.nio.file.Path;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/api")
-@Slf4j
-public class LikeController {
-    private final LikeServiceImpl likeServiceImpl;
-    @GetMapping("/like/{postId}")
-    public ResponseEntity<LikeStatus> changeLikeStatus(@PathVariable(name = "postId")Long postId){
-        return ResponseEntity.ok(likeServiceImpl.changeLikeStatus(postId));
-    }
-}
